
\name{compile_metric_registry}
\alias{compile_metric_registry}
%- Also NEED an '\alias' for EACH other topic documented here.
\title{Reads the metrics and units registry}
\description{
Reads the metrics and units registry from the specified file path.
}
\usage{
compile_metric_registry(registryPath)
}
%- maybe also 'usage' for other objects documented here.
\arguments{
  \item{registryPath}{
The file path to the metrics and units registry. Must be a CSV file.
}
}

\details{
}
\value{
The metrics and units registry as a data frame.
}
\references{
}
\author{
Tim Newbold <Tim.Newbold@unep-wcmc.org>
}
\note{
%%  ~~further notes~~
}

%% ~Make other sections like Warning with \section{Warning }{....} ~

\seealso{
%% ~~objects to See Also as \code{\link{help}}, ~~~
}
\examples{
##---- Should be DIRECTLY executable !! ----
##-- ==>  Define data, use random,
##--        or do  help(data=index)  for the standard data sets.
}