\name{Ecoregions}
\docType{data}
\alias{ecoregions}
\alias{realms}
\alias{biomes}
\alias{biomes.grid}
\title{ecoregions}
\description{Data from The Nature Conservancy's Terrestrial Ecoregions of 
the world GIS layer.}
\usage{realms
biomes
biomes.grid
}
\format{biomes.grid is a \code{SpatialPolygonsDataFrame}.

Other objects are \code{data.frame}s.}
\source{TNC. Blah.}
\references{Blah}
\keyword{datasets}
